inspect charge, made credits decimal instead of integer

+++ database/migrations/2019_01_21_212318_create_airtimes_table.php
@@ -18,7 +18,8 @@ class CreateAirtimesTable extends Migration
 		Schema::create('airtimes', function(Blueprint $table) {
             $table->increments('id');
             $table->string('key')->unique();
-            $table->integer('credits');
+//            $table->integer('credits');
+            $table->decimal('credits');
             $table->timestamps();
 		});
 	}
